requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
pypdf>=5.0.0
faiss-cpu>=1.1.0
sentence-transformers>=3.0.0
google-generativeai>=0.8.0
streamlit>=1.38.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
